fix: dario merge

+++ app/src/main/java/de/fhe/ai/colivingpilot/storage/Repository.kt
@@ -1,26 +1,72 @@
 package de.fhe.ai.colivingpilot.storage
 
+import android.content.Context
+import dagger.hilt.android.qualifiers.ApplicationContext
+import android.util.Log
 import de.fhe.ai.colivingpilot.core.CoLiPiApplication
+import de.fhe.ai.colivingpilot.network.RetrofitClient
 import de.fhe.ai.colivingpilot.model.ShoppingListItem
 import de.fhe.ai.colivingpilot.model.Task
 import de.fhe.ai.colivingpilot.model.User
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.withContext
+import org.json.JSONObject
+import java.lang.Exception
 
 
-class Repository(
-
-) {
+class Repository {
     private val db: WgDatabase = WgDatabase.getInstance(CoLiPiApplication.applicationContext())
     private val userDao: UserDao = db.userDao()
     private val taskDao: TaskDao = db.taskDao()
     private val shoppingListItemDao: ShoppingListItemDao = db.shoppingListItemDao()
     private val taskAssignedUserDao: TaskAssignedUserDao = db.taskAssignedUserDao()
 
-    suspend fun updateAll() = withContext(Dispatchers.IO) {
-        userDao.deleteAll() // triggers cascading deletes in all tables
-        // TODO: ... get json data from backend and insert everything
+    /**
+     * Refreshes the app's data by fetching the relevant dataset from the server.
+     *
+     * @return A pair indicating the success or failure of the operation and an optional error message.
+     */
+    suspend fun refresh(): Pair<Boolean, String> = withContext(Dispatchers.IO) {
+        try {
+            val response = RetrofitClient.instance.getWgData().execute()
+            if (!response.isSuccessful) {
+                val body = response.errorBody()
+                body?.let {
+                    val json = JSONObject(body.string())
+                    return@withContext Pair(false, json.getString("status"))
+                }
+                return@withContext Pair(false, "")
+            }
+
+            CoLiPiApplication.instance.keyValueStore.writeString("wg_name", "")
+            CoLiPiApplication.instance.keyValueStore.writeString("wg_code", "")
+            CoLiPiApplication.instance.keyValueStore.writeInt("wg_max_members", -1)
+
+            userDao.deleteAll() // triggers cascading deletes in all tables
+            taskDao.deleteAll() // TODO: removable once Task has reference to User
+
+            val body = response.body()
+            body?.let { resp ->
+                CoLiPiApplication.instance.keyValueStore.writeString("wg_name", resp.data.wg.name)
+                CoLiPiApplication.instance.keyValueStore.writeString("wg_code", resp.data.wg.invitationCode)
+                CoLiPiApplication.instance.keyValueStore.writeInt("wg_max_members", resp.data.wg.maximumMembers)
+                val creator = resp.data.wg.creator
+                resp.data.wg.members.forEach { member ->
+                    db.userDao().insert(User(member.id, member.username, member.beercounter, creator.username == member.username))
+                }
+                resp.data.wg.shoppingList.forEach { item ->
+                    db.shoppingListItemDao().insert(ShoppingListItem(item.id, item.title, item.notes, item.creator.id, item.isChecked))
+                }
+                resp.data.wg.tasks.forEach { task ->
+                    db.taskDao().upsert(Task(task.id, task.title, task.description, task.beerbonus))
+                }
+            }
+            return@withContext Pair(true, "")
+        } catch (_: Exception) {
+            Log.e(CoLiPiApplication.LOG_TAG, "Failed to fetch WG data")
+            return@withContext Pair(false, "")
+        }
     }
 
     fun getUsersFlow(): Flow<List<User>> {
@@ -47,8 +93,25 @@ class Repository(
         return taskDao.getTask(id)
     }
 
-    fun getShoppingListItems(): List<ShoppingListItem> {
-        return shoppingListItemDao.getShoppingListItems()
+    fun getShoppingListItemsFlow(): Flow<List<ShoppingListItem>> {
+        return shoppingListItemDao.getShoppingListItemsFlow()
     }
 
+    fun deleteItemFromShoppingList(shoppingListItem: ShoppingListItem){
+        shoppingListItemDao.deleteItemFromShoppingList(shoppingListItem)
+    }
+
+    fun insertShoppingListItem(shoppingListItem: ShoppingListItem){
+        shoppingListItemDao.insert(shoppingListItem)
+    }
+
+    fun updateItem(shoppingListItem: ShoppingListItem, boolean: Boolean){
+        var updatedItem = shoppingListItem.copy(isChecked = boolean)
+        shoppingListItemDao.updateItem(updatedItem)
+    }
+
+    //TODO später kommt der User aus der Anmeldung
+    fun getTestUser(): User {
+        return userDao.getTestUser()
+    }
 }
